Revert "feat(vr): add OpenComposite menu support (#1880)"
This reverts commit cf28c87d9c1cb42af92142751f7ce1eced7caa4d.
