+++ src/Menu.cpp
@@ -232,6 +232,8 @@ Menu::~Menu()
 	ImGui_ImplWin32_Shutdown();
 	ImGui::DestroyContext();
 	dxgiAdapter3 = nullptr;
+
+	globals::features::vr.DestroyOverlay();
 }
 
 void Menu::Load(json& o_json)
@@ -648,6 +650,10 @@ void Menu::Init()
 	}
 
 	BuildCategoryCounts();
+
+	if (globals::features::vr.IsOpenVRCompatible()) {
+		globals::features::vr.EnsureOverlayInitialized();
+	}
 
 	initialized = true;
 }
